pyproject.toml
[tool.poetry]
name = "range-monitoring-agent"
version = "0.1.0"
description = "Lightweight agent that monitors product range performance across retail store locations"
authors = []
readme = "README.md"
packages = [{include = "range_monitor"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0"
pyyaml = "^6.0"
openpyxl = "^3.1"
tabulate = "^0.9"
streamlit = "^1.35"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.4"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-v", "--import-mode=importlib"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
